refactor `ovirtapi` to separate concerns
`ovirtapi` has been split to have transport, data transform and
API interactions in separate files:

 - HTTP transport functions are in `ovirtapi/transport.js`

 - data transform functions (from API to internal structures and
   back again) are in `ovirtapi/transform.js`

 - API interaction functions are in `ovirtapi/index.js`

 - `ovirtapi/index.js` presents the save facade as the previous
   `ovirtapi.js`
