+++ config/webpack.config.dev.js
@@ -73,7 +73,7 @@ module.exports = {
       // Support React Native Web
       // https://www.smashingmagazine.com/2016/08/a-glimpse-into-the-future-with-react-native-for-web/
       'react-native': 'react-native-web',
-      'ovirtapi': './ovirtapi',
+      'ovirtapi': `${paths.appSrc}/ovirtapi/index.js`,
     },
   },
   // Resolve loaders (webpack plugins for CSS, images, transpilation) from the
